renamed coincident to state_constraints and refactored to_part_file to use single dispatch for constraints in a separate sketch_constraints file
