+++ src/PanCAD/geometry/constraints/__init__.py
@@ -1,4 +1,6 @@
-from PanCAD.geometry.constraints.coincident import Coincident
+from PanCAD.geometry.constraints.state_constraint import (
+    Coincident, Equal, Parallel, Perpendicular, Tangent
+)
 from PanCAD.geometry.constraints.snapto import Horizontal, Vertical
 from PanCAD.geometry.constraints.distance import (
     HorizontalDistance, VerticalDistance, Distance, Radius, Diameter,
